fix(windows): two more /W4 /WX diagnostics, and -k earns its keep
The Ninja + -k switch worked exactly as intended: this one push
surfaced two distinct root causes instead of one, spread across 10
build failures.

- CommandSerializer.h:10 (C4245, signed/unsigned mismatch): the
  ASM_LEVEL_DISABLED sentinel was `unsigned int x = -1`. Same 10
  failures were really this one header, hit by everything that
  includes it (sony-transport itself, and 9 protocol/legacy test
  translation units). static_cast<unsigned int>(-1) makes the
  bit-pattern-preserving intent explicit.
- IpcProtocol.cpp (C4244, via std::transform -> std::tolower):
  std::tolower returns int; writing it into a std::string via
  std::transform narrows silently. static_cast<char> on the lambda's
  return makes it explicit. GCC/Clang don't warn on this extremely
  common idiom; MSVC's /W4 does.

+++ libs/sony-transport/legacy/CommandSerializer.h
@@ -7,7 +7,7 @@
 #include "Exceptions.h"
 
 constexpr int MINIMUM_VOICE_FOCUS_STEP = 2;
-constexpr unsigned int ASM_LEVEL_DISABLED = -1;
+constexpr unsigned int ASM_LEVEL_DISABLED = static_cast<unsigned int>(-1);
 
 namespace CommandSerializer
 {
